+++ src/StreamStore.Contracts/Database/IStreamDatabase.cs
@@ -5,7 +5,7 @@ namespace StreamStore
 {
     public interface IStreamDatabase: IStreamReader
     {
-        Task<EventMetadataRecordCollection?> FindMetadataAsync(Id streamId, CancellationToken token = default);
+        Task<Revision?> GetActualRevision(Id streamId, CancellationToken token = default);
         Task DeleteAsync(Id streamId, CancellationToken token = default);
         Task<IStreamUnitOfWork> BeginAppendAsync(Id streamId, Revision expectedStreamVersion, CancellationToken token = default);
     }