+++ src/StreamStore.NoSql.Cassandra/Database/CassandraCqlQueries.cs
@@ -21,11 +21,6 @@ namespace StreamStore.NoSql.Cassandra.Database
             return new Cql($"DELETE FROM {config.EventsTableName} WHERE stream_id = ?", streamId);
         }
 
-        public Cql StreamMetadata(string streamId)
-        {
-            return new Cql($"SELECT id, stream_id, timestamp, revision  FROM {config.EventsTableName} WHERE stream_id = ?", streamId);
-        }
-
         public Cql StreamEvents(string streamId, int from, int count)
         {
             return new Cql($"SELECT id, stream_id, timestamp, revision, data FROM {config.EventsTableName} WHERE stream_id = ? AND revision >= ? LIMIT ?", streamId, from, count);